Compile libsurvive in release mode to suppress asserts (#6)
It seems like there are some aggressive asserts in libsurvive which end up sporadically halting the driver application from running. To avoid these, compile libsurvive in Release mode. Also, add gdb and valgrind to the docker image to facilitate easier debugging, and fix the /dev cgroup permissions to prevent having to run in privileged mode -- we really shouldn't need this for a simple tracking app.